修正属性读取的精度问题, 以后干脆全用浮点算了

+++ Samples/Frontline.GameDesign/Designs/兵种/DEquip.cs
@@ -12,8 +12,8 @@ namespace Frontline.GameDesign
         public int type { get; set; }
         public int pos { get; set; }
         public int base_attr_type { get; set; }
-        public int base_attr_value { get; set; }
-        public int level_grow { get; set; }
+        public float base_attr_value { get; set; }
+        public float level_grow { get; set; }
         public int level_k { get; set; }
         public int gradeid { get; set; }
     }
